Updated importing, more efficient.

+++ application/controllers/ImportController.php
@@ -13,84 +13,28 @@ class ImportController extends Zend_Controller_Action
         // action body
     }
     
-    public function importPlayerListAction()
+        
+    public function importTownsAction()
     {
-    	$mtime = microtime(true);
-    	$n = 0;
-    	$link = 'http://en56.grepolis.com/data/players.txt.gz';
     	$config = Zend_Registry::get('config');
-    	if(!is_dir($config->serverdata->path."EN56/")){
-    		mkdir($config->serverdata->path."EN56/");
-    	}
-    	$filename = $config->serverdata->path."EN56/players.txt.".time().".gz";
-    	file_put_contents($filename,file_get_contents($link));
-    	$lines = gzfile($filename);    	 
-    	foreach($lines as $line){
-    		$n++;
-    		$data = explode(',',$line);
-    		$existing = Crumblez_Model_Player::getObjectId((int)trim($data[0]));
-    		if($existing)$player = Crumblez_Model_Player::get((int)$existing);
-    		else $player = new Crumblez_Model_Player(array());
-    		$player->playerId = trim($data[0]);
-    		$player->name = str_replace('+',' ',trim($data[1]));
-    		$player->allianceId = trim($data[2]);
-    		$player->points = trim($data[3]);
-    		$player->rank = trim($data[4]);
-    		$player->towns = trim($data[5]);
-    		$player->save();
-    	}
-    	 
-    	die("Updated ".$n." players. It took ".(microtime(true)-$mtime)." seconds");
-    }
-    
-    public function importAllianceListAction()
-    {
-    	$mtime = microtime(true);
-    	$n = 0;
-    	$link = 'http://en56.grepolis.com/data/alliances.txt.gz';
-    	$config = Zend_Registry::get('config');
-    	if(!is_dir($config->serverdata->path."EN56/")){
-    		mkdir($config->serverdata->path."EN56/");
-    	}
-    	$filename = $config->serverdata->path."EN56/alliances.txt.".time().".gz";
-    	file_put_contents($filename,file_get_contents($link));
-    	$lines = gzfile($filename);    	 
-    	foreach($lines as $line){
-    		$n++;
-    		$data = explode(',',$line);
-    		$existing = Crumblez_Model_Alliance::getObjectId((int)trim($data[0]));
-    		if($existing)$alliance = Crumblez_Model_Alliance::get((int)$existing);
-    		else $alliance = new Crumblez_Model_Alliance(array());
-    		$alliance->allianceId = trim($data[0]);
-    		$alliance->name = str_replace('+',' ',trim($data[1]));
-    		$alliance->points = trim($data[2]);
-    		$alliance->towns = trim($data[3]);
-    		$alliance->members = trim($data[4]);
-    		$alliance->rank = trim($data[5]);
-    		$alliance->save();
-    	}
-    	 
-    	die("Updated ".$n." alliances. It took ".(microtime(true)-$mtime)." seconds");
-    }
-    
-    public function importTownListAction()
-    {
+    	if($this->_getParam('cronkey') != $config->cronkey || !$this->_getParam('cronkey')) throw new Zend_Controller_Action_Exception('This page dont exist',404);
     	$mtime = microtime(true);
     	$n = 0;
     	$link = 'http://en56.grepolis.com/data/towns.txt.gz';
-    	$config = Zend_Registry::get('config');
     	if(!is_dir($config->serverdata->path."EN56/")){
     		mkdir($config->serverdata->path."EN56/");
     	}
     	$filename = $config->serverdata->path."EN56/towns.txt.".time().".gz";
     	file_put_contents($filename,file_get_contents($link));
     	$lines = gzfile($filename);
+    	unlink($filename);
     	foreach($lines as $line){
     		$n++;
     		$data = explode(',',$line);
     		$existing = Crumblez_Model_Town::getObjectId((int)trim($data[0]));
     		if($existing)$town = Crumblez_Model_Town::get((int)$existing);
     		else $town = new Crumblez_Model_Town(array());
+    		if(!$town->getId())$town->timeFounded = time();
     		$town->townId = trim($data[0]);
     		$town->playerId = trim($data[1]);
     		$town->name = str_replace('+',' ',trim($data[2]));
@@ -105,18 +49,20 @@ class ImportController extends Zend_Controller_Action
     }
     
     
-    public function importConquerListAction()
+    public function importConquersAction()
     {
+    	$config = Zend_Registry::get('config');
+    	if($this->_getParam('cronkey') != $config->cronkey || !$this->_getParam('cronkey')) throw new Zend_Controller_Action_Exception('This page dont exist',404);
     	$mtime = microtime(true);
     	$n = 0;
     	$link = 'http://en56.grepolis.com/data/conquers.txt.gz';
-    	$config = Zend_Registry::get('config');
     	if(!is_dir($config->serverdata->path."EN56/")){
     		mkdir($config->serverdata->path."EN56/");
     	}
     	$filename = $config->serverdata->path."EN56/conquers.txt.".time().".gz";
     	file_put_contents($filename,file_get_contents($link));
     	$lines = gzfile($filename);
+    	unlink($filename);
     	Crumblez_Model_Conquer::clearTable();
     	foreach($lines as $line){
     		$n++;
@@ -137,11 +83,12 @@ class ImportController extends Zend_Controller_Action
     
     public function importIslandListAction()
     {
+    	$config = Zend_Registry::get('config');
+    	if($this->_getParam('cronkey') != $config->cronkey || !$this->_getParam('cronkey')) throw new Zend_Controller_Action_Exception('This page dont exist',404);
     	$mtime = microtime(true);
 	    $n = 0;
     	if(!Crumblez_Model_Island::chkStatus()){//if islands already exist, don't download again.
 	    	$link = 'http://en56.grepolis.com/data/islands.txt.gz';
-	    	$config = Zend_Registry::get('config');
 	    	if(!is_dir($config->serverdata->path."EN56/")){
 	    		mkdir($config->serverdata->path."EN56/");
 	    	}
@@ -163,70 +110,19 @@ class ImportController extends Zend_Controller_Action
     	die("Updated ".$n." islands. It took ".(microtime(true)-$mtime)." seconds");
     }
     
-    public function importPlayerBattlePointListAction()
+    
+    public function importAllianceAction()
     {
-    	$mtime = microtime(true);
-    	$n = 0;
-    	$bpTotalLink = 'http://en56.grepolis.com/data/player_kills_all.txt.gz';
-    	$bpALink = 'http://en56.grepolis.com/data/player_kills_att.txt.gz';
-    	$bpDLink = 'http://en56.grepolis.com/data/player_kills_def.txt.gz';
-    	 
     	$config = Zend_Registry::get('config');
-    	if(!is_dir($config->serverdata->path."EN56/")){
-    		mkdir($config->serverdata->path."EN56/");
-    	}
-    	$filenameT = $config->serverdata->path."EN56/player_kills_all.txt.".time().".gz";
-    	file_put_contents($filenameT,file_get_contents($bpTotalLink));
-    	$filenameA = $config->serverdata->path."EN56/player_kills_att.txt.".time().".gz";
-    	file_put_contents($filenameA,file_get_contents($bpALink));
-    	$filenameD = $config->serverdata->path."EN56/player_kills_def.txt.".time().".gz";
-    	file_put_contents($filenameD,file_get_contents($bpDLink));
-    	$linesT = gzfile($filenameT);
-    	$linesA = gzfile($filenameA);
-    	$linesD = gzfile($filenameD);
-    	$BPArray = array();
-    	foreach($linesT as $line){
-    		$data = explode(',',$line);
-    		$BPArray[$data[1]]['bp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['bp'] = trim($data[2]);
-    	}
-    	foreach($linesA as $line){
-    		$data = explode(',',$line);
-    		$BPArray[$data[1]]['abp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['abp'] = trim($data[2]);
-    	}
-    	foreach($linesD as $line){
-    		$data = explode(',',$line);
-    		$BPArray[$data[1]]['dbp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['dbp'] = trim($data[2]);
-    	}
-    	 
-    	foreach($BPArray as $playerId=>$data){
-    		$n++;
-    		$battlePointsPlayer = new Crumblez_Model_BattlePointsPlayer(array());
-    		$battlePointsPlayer->playerId = $playerId;
-    		$battlePointsPlayer->time = time();
-    		$battlePointsPlayer->points = $data['bp'];
-    		$battlePointsPlayer->pointsAttack = $data['abp'];
-    		$battlePointsPlayer->pointsDefense = $data['dbp'];
-    		$battlePointsPlayer->pointsRank = $data['bp_rank'];
-    		$battlePointsPlayer->pointsAttackRank = $data['abp_rank'];
-    		$battlePointsPlayer->pointsDefenseRank = $data['dbp_rank'];
-    		$battlePointsPlayer->save();
-    	}
-    
-    	die("Updated ".$n." Player BP listings. It took ".(microtime(true)-$mtime)." seconds");
-    }
-    
-    public function importAllianceBattlePointListAction()
-    {
+    	if($this->_getParam('cronkey') != $config->cronkey || !$this->_getParam('cronkey')) throw new Zend_Controller_Action_Exception('This page dont exist',404);
     	$mtime = microtime(true);
     	$n = 0;
     	$bpTotalLink = 'http://en56.grepolis.com/data/alliance_kills_all.txt.gz';
     	$bpALink = 'http://en56.grepolis.com/data/alliance_kills_att.txt.gz';
     	$bpDLink = 'http://en56.grepolis.com/data/alliance_kills_def.txt.gz';
-    	 
-    	$config = Zend_Registry::get('config');
+    	$link = 'http://en56.grepolis.com/data/alliances.txt.gz';
+    	$time = time();
+    	
     	if(!is_dir($config->serverdata->path."EN56/")){
     		mkdir($config->serverdata->path."EN56/");
     	}
@@ -236,41 +132,176 @@ class ImportController extends Zend_Controller_Action
     	file_put_contents($filenameA,file_get_contents($bpALink));
     	$filenameD = $config->serverdata->path."EN56/alliance_kills_def.txt.".time().".gz";
     	file_put_contents($filenameD,file_get_contents($bpDLink));
+    	$filename = $config->serverdata->path."EN56/alliances.txt.".time().".gz";
+    	file_put_contents($filename,file_get_contents($link));
     	$linesT = gzfile($filenameT);
     	$linesA = gzfile($filenameA);
     	$linesD = gzfile($filenameD);
-    	$BPArray = array();
+    	$lines = gzfile($filename);
+    	unlink($filenameT);
+    	unlink($filenameA);
+    	unlink($filenameD);
+    	unlink($filename);
+    	$AllianceArray = array();
     	foreach($linesT as $line){
     		$data = explode(',',$line);
-    		$BPArray[$data[1]]['bp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['bp'] = trim($data[2]);
+    		$AllianceArray[$data[1]]['bp_rank'] = trim($data[0]);
+    		$AllianceArray[$data[1]]['bp'] = trim($data[2]);
     	}
     	foreach($linesA as $line){
     		$data = explode(',',$line);
-    		$BPArray[$data[1]]['abp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['abp'] = trim($data[2]);
+    		$AllianceArray[$data[1]]['abp_rank'] = trim($data[0]);
+    		$AllianceArray[$data[1]]['abp'] = trim($data[2]);
     	}
     	foreach($linesD as $line){
     		$data = explode(',',$line);
-    		$BPArray[$data[1]]['dbp_rank'] = trim($data[0]);
-    		$BPArray[$data[1]]['dbp'] = trim($data[2]);
+    		$AllianceArray[$data[1]]['dbp_rank'] = trim($data[0]);
+    		$AllianceArray[$data[1]]['dbp'] = trim($data[2]);
     	}
-    	 
-    	foreach($BPArray as $allianceId=>$data){
+    	
+    	foreach($lines as $line){
+    		$data = explode(',',$line);
+    		$AllianceArray[$data[0]]['name'] = str_replace('+',' ',trim($data[1]));
+    		$AllianceArray[$data[0]]['points'] = trim($data[2]);
+    		$AllianceArray[$data[0]]['towns'] = trim($data[3]);
+    		$AllianceArray[$data[0]]['members'] = trim($data[4]);
+    		$AllianceArray[$data[0]]['rank'] = trim($data[5]);
+    	}
+    	    	
+    	foreach($AllianceArray as $allianceId=>$data){
+    		$allianceLatest = Crumblez_Model_Alliance::getLatestAllianceData($allianceId);
+    		if(
+    				$allianceLatest->name==$data['name'] &&
+    				$allianceLatest->towns==$data['towns'] &&
+    				$allianceLatest->members==$data['members'] &&
+    				$allianceLatest->points==$data['points'] &&
+    				$allianceLatest->pointsBattle==$data['bp'] &&
+    				$allianceLatest->pointsAttack==$data['abp'] &&
+    				$allianceLatest->pointsDefense==$data['dbp'] &&
+    				$allianceLatest->pointsRank==$data['rank'] &&
+    				$allianceLatest->pointsBattleRank==$data['bp_rank'] &&
+    				$allianceLatest->pointsAttackRank==$data['abp_rank'] &&
+    				$allianceLatest->pointsDefenseRank==$data['dbp_rank']
+    		){
+    			continue;
+    		}
     		$n++;
-    		$battlePointsPlayer = new Crumblez_Model_BattlePointsAlliance(array());
-    		$battlePointsPlayer->allianceId = $allianceId;
-    		$battlePointsPlayer->time = time();
-    		$battlePointsPlayer->points = $data['bp'];
-    		$battlePointsPlayer->pointsAttack = $data['abp'];
-    		$battlePointsPlayer->pointsDefense = $data['dbp'];
-    		$battlePointsPlayer->pointsRank = $data['bp_rank'];
-    		$battlePointsPlayer->pointsAttackRank = $data['abp_rank'];
-    		$battlePointsPlayer->pointsDefenseRank = $data['dbp_rank'];
-    		$battlePointsPlayer->save();
+    		$alliance = new Crumblez_Model_Alliance(array());
+    		$alliance->allianceId = $allianceId;
+    		$alliance->time = $time;
+    		$alliance->name = $data['name'];
+    		$alliance->points = $data['points'];
+    		$alliance->towns = $data['towns'];
+    		$alliance->members = $data['members'];
+    		$alliance->points = $data['points'];
+    		$alliance->pointsBattle = $data['bp'];
+    		$alliance->pointsAttack = $data['abp'];
+    		$alliance->pointsDefense = $data['dbp'];
+    		$alliance->pointsRank = $data['rank'];
+    		$alliance->pointsBattleRank = $data['bp_rank'];
+    		$alliance->pointsAttackRank = $data['abp_rank'];
+    		$alliance->pointsDefenseRank = $data['dbp_rank'];
+    		$alliance->save();
     	}
     
-    	die("Updated ".$n." Alliance BP listings. It took ".(microtime(true)-$mtime)." seconds");
+    	die("Updated ".$n." Alliance listings. It took ".(microtime(true)-$mtime)." seconds");
+    }
+    
+    
+    public function importPlayerAction()
+    {
+    	$config = Zend_Registry::get('config');
+    	if($this->_getParam('cronkey') != $config->cronkey || !$this->_getParam('cronkey')) throw new Zend_Controller_Action_Exception('This page dont exist',404);
+    	$mtime = microtime(true);
+    	$n = 0;
+    	$bpTotalLink = 'http://en56.grepolis.com/data/player_kills_all.txt.gz';
+    	$bpALink = 'http://en56.grepolis.com/data/player_kills_att.txt.gz';
+    	$bpDLink = 'http://en56.grepolis.com/data/player_kills_def.txt.gz';
+    	$link = 'http://en56.grepolis.com/data/players.txt.gz';
+    	$time = time();
+    	
+    	if(!is_dir($config->serverdata->path."EN56/")){
+    		mkdir($config->serverdata->path."EN56/");
+    	}
+    	$filenameT = $config->serverdata->path."EN56/player_kills_all.txt.".time().".gz";
+    	file_put_contents($filenameT,file_get_contents($bpTotalLink));
+    	$filenameA = $config->serverdata->path."EN56/player_kills_att.txt.".time().".gz";
+    	file_put_contents($filenameA,file_get_contents($bpALink));
+    	$filenameD = $config->serverdata->path."EN56/player_kills_def.txt.".time().".gz";
+    	file_put_contents($filenameD,file_get_contents($bpDLink));
+    	$filename = $config->serverdata->path."EN56/players.txt.".time().".gz";
+    	file_put_contents($filename,file_get_contents($link));
+    	$linesT = gzfile($filenameT);
+    	$linesA = gzfile($filenameA);
+    	$linesD = gzfile($filenameD);
+    	$lines = gzfile($filename);
+    	unlink($filenameT);
+    	unlink($filenameA);
+    	unlink($filenameD);
+    	unlink($filename);
+    	$PlayerArray = array();
+    	foreach($linesT as $line){
+    		$data = explode(',',$line);
+    		$PlayerArray[$data[1]]['bp_rank'] = trim($data[0]);
+    		$PlayerArray[$data[1]]['bp'] = trim($data[2]);
+    	}
+    	foreach($linesA as $line){
+    		$data = explode(',',$line);
+    		$PlayerArray[$data[1]]['abp_rank'] = trim($data[0]);
+    		$PlayerArray[$data[1]]['abp'] = trim($data[2]);
+    	}
+    	foreach($linesD as $line){
+    		$data = explode(',',$line);
+    		$PlayerArray[$data[1]]['dbp_rank'] = trim($data[0]);
+    		$PlayerArray[$data[1]]['dbp'] = trim($data[2]);
+    	}
+    	foreach($lines as $line){
+    		$data = explode(',',$line);
+    		$PlayerArray[$data[0]]['name'] = str_replace('+',' ',trim($data[1]));
+    		$PlayerArray[$data[0]]['allianceId'] = $data[2]?trim($data[2]):null;
+    		$PlayerArray[$data[0]]['points'] = trim($data[3]);
+    		$PlayerArray[$data[0]]['rank'] = trim($data[4]);
+    		$PlayerArray[$data[0]]['towns'] = trim($data[5]);
+    	}
+    		
+    	
+    	foreach($PlayerArray as $playerId=>$data){
+    		$playerLatest = Crumblez_Model_Player::getLatestPlayerData($playerId);
+    		if(
+    				$playerLatest->name==$data['name'] &&
+    				$playerLatest->allianceId==$data['allianceId'] &&
+    				$playerLatest->towns==$data['towns'] &&
+    				$playerLatest->points==$data['points'] &&
+    				$playerLatest->pointsBattle==$data['bp'] &&
+    				$playerLatest->pointsAttack==$data['abp'] &&
+    				$playerLatest->pointsDefense==$data['dbp'] &&
+    				$playerLatest->pointsRank==$data['rank'] &&
+    				$playerLatest->pointsBattleRank==$data['bp_rank'] &&
+    				$playerLatest->pointsAttackRank==$data['abp_rank'] &&
+    				$playerLatest->pointsDefenseRank==$data['dbp_rank']
+    		){
+    			continue;
+    		}
+    		$n++;
+    		$player = new Crumblez_Model_Player(array());
+    		$player->playerId = $playerId;
+    		$player->allianceId = $data['allianceId'];
+    		$player->time = $time;
+    		$player->name = $data['name'];
+    		$player->points = $data['points'];
+    		$player->towns = $data['towns'];
+    		$player->points = $data['points'];
+    		$player->pointsBattle = $data['bp'];
+    		$player->pointsAttack = $data['abp'];
+    		$player->pointsDefense = $data['dbp'];
+    		$player->pointsRank = $data['rank'];
+    		$player->pointsBattleRank = $data['bp_rank'];
+    		$player->pointsAttackRank = $data['abp_rank'];
+    		$player->pointsDefenseRank = $data['dbp_rank'];
+    		$player->save();
+    	}
+    
+    	die("Updated ".$n." Player listings. It took ".(microtime(true)-$mtime)." seconds");
     }
     
 }
